requirements.txt
mcp>=0.9.0
python-gitlab>=4.0.0

